separate message-header own component
